move round related functions to round.js
